export Composable fun as class -- prepare for CMP
